Bring back ReduxProvider
