+++ packages/web/src/test/test-utils.tsx
@@ -2,18 +2,40 @@ import { ReactElement, ReactNode } from 'react'
 
 import { ThemeProvider } from '@audius/harmony'
 import { render, RenderOptions } from '@testing-library/react'
+import { Provider } from 'react-redux'
 import { Router } from 'react-router-dom'
 import { CompatRouter } from 'react-router-dom-v5-compat'
 import { PartialDeep } from 'type-fest'
 
-import { HistoryContext, HistoryContextProvider } from 'app/HistoryProvider'
-import { ReduxProvider } from 'app/ReduxProvider'
+import {
+  HistoryContext,
+  HistoryContextProvider,
+  useHistoryContext
+} from 'app/HistoryProvider'
 import { RouterContextProvider } from 'components/animated-switch/RouterContextProvider'
 import { ToastContextProvider } from 'components/toast/ToastContext'
+import { useIsMobile } from 'hooks/useIsMobile'
+import { configureStore } from 'store/configureStore'
 import { AppState } from 'store/types'
 
 type TestOptions = {
   reduxState?: PartialDeep<AppState>
+}
+
+type ReduxProviderProps = {
+  children: ReactNode
+  initialStoreState?: PartialDeep<AppState>
+}
+
+export const ReduxProvider = ({
+  children,
+  initialStoreState
+}: ReduxProviderProps) => {
+  const { history } = useHistoryContext()
+  const isMobile = useIsMobile()
+  const { store } = configureStore(history, isMobile, initialStoreState)
+
+  return <Provider store={store}>{children}</Provider>
 }
 
 type TestProvidersProps = {
